backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Flask-Marshmallow==0.15.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
PyMySQL==1.1.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.8
redis==5.0.1
flasgger==0.9.7.1
gunicorn==21.2.0
gevent==23.9.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
